refactored changelog
