+++ apps/devportal/src/components/changelog/ChangelogFilter.tsx
@@ -1,5 +1,5 @@
-import Product from '@/../../packages/sc-changelog/types/product';
 import { useState } from 'react';
+import Product from 'sc-changelog/types/product';
 import MultiSelect, { Option } from 'ui/components/dropdown/MultiSelect';
 
 type ChangelogFilterProps = {
