+++ apps/devportal/src/pages/api/changelog/all/index.ts
@@ -1,12 +1,10 @@
 // Interfaces
-import { getOverviewPerMonth } from '@/src/common/changelog';
 import type { NextApiRequest, NextApiResponse } from 'next';
+import { GetSummaryLatestItemsByProductAndChangeType } from 'sc-changelog/changelog';
 import { ChangelogEntrySummary } from 'sc-changelog/types/changeLogEntry';
-
-const getQueryArray = (query: string | string[] | undefined): string[] => {
-  if (query == undefined) return [];
-  return Array.isArray(query) ? query : [query];
-};
+import ChangeType from 'sc-changelog/types/changeType';
+import Product from 'sc-changelog/types/product';
+import { getQueryArray } from 'sc-changelog/utils/requests';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse<Record<string, ChangelogEntrySummary[]>>) => {
   const products: string[] = getQueryArray(req.query.product);
@@ -18,3 +16,32 @@ const handler = async (req: NextApiRequest, res: NextApiResponse<Record<string, 
 };
 
 export default handler;
+
+const getOverviewPerMonth: any = async (isPreview: boolean, products?: Product[], changes?: ChangeType[]) => {
+  const items = await GetSummaryLatestItemsByProductAndChangeType(isPreview, products?.join('|'), changes?.join('|'));
+  const entries: ChangelogEntrySummary[] = items.entries;
+
+  // Group the entries by month
+  const groupedObjects = entries.reduce((collection, obj) => {
+    const monthYear = new Date(obj.releaseDate).toLocaleString('en-US', { month: 'short', year: 'numeric' });
+    if (!collection[monthYear]) {
+      collection[monthYear] = [];
+    }
+    collection[monthYear].push(obj);
+
+    // Sort updates within a month (latest first)
+    collection[monthYear].sort((a, b) => {
+      const earliestDateA = new Date(a.releaseDate);
+      const earliestDateB = new Date(b.releaseDate);
+      return earliestDateB.getTime() - earliestDateA.getTime();
+    });
+    return collection;
+  }, {} as Record<string, ChangelogEntrySummary[]>);
+
+  // Sort the keys (year-month)
+  const sorted = Object.entries(groupedObjects)
+    .sort(([dateA], [dateB]) => new Date(dateB).getTime() - new Date(dateA).getTime())
+    .reduce((acc, [date, value]) => ({ ...acc, [date]: value }), {});
+
+  return sorted;
+};
