+++ apps/devportal/src/pages/api/changelog/products/index.ts
@@ -2,16 +2,11 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { GetProducts } from 'sc-changelog/changelog';
 import Product from 'sc-changelog/types/product';
-
-const getQueryValue = (query: string | string[] | undefined): string => {
-  if (query == undefined) return '';
-
-  return Array.isArray(query) ? query[0] : query;
-};
+import { getQueryValue } from 'sc-changelog/utils/requests';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse<Product[]>) => {
   const showAll: boolean = getQueryValue(req.query.all) == 'false' ? false : true;
-  res.setHeader('Cache-Control', 's-maxage=600, stale-while-revalidate');
+  res.setHeader('Cache-Control', 'stale-while-revalidate');
 
   await GetProducts().then((response: Product[]) => {
     if (showAll) {
