+++ apps/devportal/src/pages/api/changelog/types/index.ts
@@ -1,11 +1,11 @@
 // Interfaces
-import { GetChangeTypes } from '@/../../packages/sc-changelog/changelog';
-import ChangeType from '@/../../packages/sc-changelog/types/changeType';
 import type { NextApiRequest, NextApiResponse } from 'next';
+import { GetChangeTypes } from 'sc-changelog/changelog';
+import ChangeType from 'sc-changelog/types/changeType';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse<ChangeType[]>) => {
   await GetChangeTypes().then((response: ChangeType[]) => {
-    res.setHeader('Cache-Control', 's-maxage=600, stale-while-revalidate');
+    res.setHeader('Cache-Control', 'stale-while-revalidate');
     res.status(200).json(response);
   });
 };
