+++ packages/sc-changelog/types/index.ts
@@ -19,6 +19,5 @@ export enum ProductName {
 
 export * from './changeType';
 export * from './common/media';
-export * from './releaseNotes';
 export * from './sitecoreCloud';
 export * from './sitecoreProduct';
